requirements.txt
snowflake-connector-python
python-dotenv
dbt-snowflake
